weechat

+++ Feral Wiki/Software/Weechat - IRC client basic setup/scripts/weechat.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # weechat installation
-scriptversion="1.0.2"
+scriptversion="1.0.3"
 scriptname="install.weechat"
 # randomessence
 #
@@ -95,18 +95,23 @@ then
 ####### Script Start #######
 ############################
 #
-	wget -qO ~/weechat.tar.gz "$weechat"
-	tar xf ~/weechat.tar.gz
-	cd ~/weechat-"$weechatfv"
-	sed -i 's/SET(CMAKE_SKIP_RPATH ON)//g' ~/weechat-"$weechatfv"/CMakeLists.txt
-	~/bin/cmake -DCMAKE_INSTALL_RPATH=/opt/curl/current/lib -DPREFIX="$HOME" -DCURL_LIBRARY=/opt/curl/current/lib/libcurl.so -DCURL_INCLUDE_DIR=/opt/curl/current/include
-	make
-	make install
-	cd
-	rm -rf ~/weechat.tar.gz ~/weechat-"$weechatfv"
-	echo
-	echo "Done. Continue with the rest of the FAQ to configure weechat"
-	echo
+	if [[ -f "$HOME/bin/cmake" ]]
+	then
+		wget -qO ~/weechat.tar.gz "$weechat"
+		tar xf ~/weechat.tar.gz
+		cd ~/weechat-"$weechatfv"
+		sed -i 's/SET(CMAKE_SKIP_RPATH ON)//g' ~/weechat-"$weechatfv"/CMakeLists.txt
+		~/bin/cmake -DCMAKE_INSTALL_RPATH=/opt/curl/current/lib -DPREFIX="$HOME" -DCURL_LIBRARY=/opt/curl/current/lib/libcurl.so -DCURL_INCLUDE_DIR=/opt/curl/current/include
+		make
+		make install
+		cd
+		rm -rf ~/weechat.tar.gz ~/weechat-"$weechatfv"
+		echo
+		echo "Done. Continue with the rest of the FAQ to configure weechat"
+		echo
+	else
+		echo "Cmake is not installed. Please Complete Step 1 of the FAQ before running this script"
+	fi
 #
 ############################
 ####### Script Ends  #######
